+++ src/Public/Add-BTImage.ps1
@@ -1,14 +1,20 @@
 function Add-BTImage {
     <#
         .SYNOPSIS
-        Add an image to the body of a toaster notification.
+        Add an image to a toast notification.
 
         .DESCRIPTION
-        The Add-BTImage function adds an image to the body of a toast notification.
+        The Add-BTImage function adds one of three image types to a toast notification.
 
-        Multiple images can be added by calling the function multiple times. The images are displayed in the order in which they are added.
+        Multiple images can be added to the body of a toast notification by calling the function without the AppLogo or HeroImage switches multiple times. These images are displayed in the order in which they are added.
 
         Any images added to the body of a toast notification are displayed after the text of the toast notification.
+
+        By using the AppLogo switch, this function specifies an image to be displayed on a toast notification as the app logo
+
+        Prior to Windows 10 20H2 this app logo would replace the smaller icon that represents the source of a notification, but from 20H2 it is displayed in addition to the icon.
+
+        Finally, using the HeroImage switch, this function specifies an image to be displayed across the top of a toast notification.
 
         .INPUTS
         Microsoft.Toolkit.Uwp.Notifications.ToastContentBuilder. You can pipe a toast content builder object to Add-BTImage.
@@ -84,6 +90,20 @@ function Add-BTImage {
 
         The image is displayed after the text elements, despite being added first.
 
+        .EXAMPLE
+        PS C:\>$Builder = New-BTContentBuilder
+
+        PS C:\>Add-BTImage -ContentBuilder $Builder -Source 'C:\Temp\LocalImage.png' -AppLogo
+
+        This example adds a local image as the app logo on the toast notification using a toast content builder object.
+
+        .EXAMPLE
+        PS C:\>$Builder = New-BTContentBuilder
+
+        PS C:\>Add-BTImage -ContentBuilder $Builder -Source 'C:\Temp\LocalImage.png' -HeroImage
+
+        This example adds a local image as the hero image on the toast notification using a toast content builder object.
+
         .NOTES
         You can reference an image from your local computer, a network location, or the internet.
 
@@ -95,7 +115,7 @@ function Add-BTImage {
         https://docs.toastit.dev/commands/add-btimage
     #>
 
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'AdaptiveImage')]
     param (
         # The toast content builder object that represents the toast notification being constructed.
         [Parameter(Mandatory,
@@ -106,9 +126,11 @@ function Add-BTImage {
         # Online images will be downloaded and cached in the user's TEMP directory for future use.
         [Parameter(Position = 0)]
         [ValidateNotNullOrEmpty()]
-        [string] $Source,
+        [string] $Source = $Script:DefaultImage,
 
         # Specify how the image should be cropped.
+        [Parameter(ParameterSetName = 'AdaptiveImage')]
+        [Parameter(ParameterSetName = 'AppLogo')]
         [Microsoft.Toolkit.Uwp.Notifications.AdaptiveImageCrop] $Crop = 'Default',
 
         # A description of the image, for users of assistive technologies.
@@ -118,27 +140,55 @@ function Add-BTImage {
         # Used when the online resource has been updated and you need to ensure that users see the latest version.
         [switch] $IgnoreCache,
 
+        # Specify that the image should be used as the app logo.
+        [Parameter(ParameterSetName = 'AppLogo',
+                   Mandatory)]
+        [switch] $AppLogo,
+
+        # Specify that the image should be used as the hero image.
+        [Parameter(ParameterSetName = 'HeroImage',
+                   Mandatory)]
+        [switch] $HeroImage,
+
         # Returns an object that represents the item with which you're working. By default, this cmdlet doesn't generate any output.
         [switch] $PassThru
     )
 
     begin {}
     process {
-        $BTImage = [Microsoft.Toolkit.Uwp.Notifications.AdaptiveImage ]::new()
-
-        $BTImage.Source = if ($IgnoreCache) {
+        $paramSrc = if ($IgnoreCache) {
             Optimize-BTImageSource -Source $Source -ForceRefresh
         } else {
             Optimize-BTImageSource -Source $Source
         }
 
-        $BTImage.HintCrop = $Crop
+        if ($AppLogo.IsPresent) {
+            if ($AlternateText) {
+                $null = $ContentBuilder.AddAppLogoOverride($paramSrc,
+                                                           $Crop,
+                                                           $AlternateText)
+            } else {
+                $null = $ContentBuilder.AddAppLogoOverride($paramSrc,
+                                                           $Crop)
+            }
+        } elseif ($HeroImage.IsPresent) {
+            if ($AlternateText) {
+                $null = $ContentBuilder.AddHeroImage($paramSrc,
+                                                     $AlternateText)
+            } else {
+                $null = $ContentBuilder.AddHeroImage($paramSrc)
+            }
+        } else {
+            $BTImage = [Microsoft.Toolkit.Uwp.Notifications.AdaptiveImage ]::new()
+            $BTImage.Source = $paramSrc
+            $BTImage.HintCrop = $Crop
 
-        if ($AlternateText) {
-            $BTImage.AlternateText = $AlternateText
+            if ($AlternateText) {
+                $BTImage.AlternateText = $AlternateText
+            }
+
+            $null = $ContentBuilder.AddVisualChild($BTImage)
         }
-
-        $null = $ContentBuilder.AddVisualChild($BTImage)
 
         if ($PassThru) {
             $ContentBuilder